area_of_rectangle

+++ 04.Input().py/06.area_of_rectangle.py
@@ -1,5 +1,5 @@
-l = int(input("enter your length:"))
-b = int(input("enter your breadth:"))
+l = float(input("enter your length:"))
+b = float(input("enter your breadth:"))
 
 area_of_rectangle=(l*b)
 print("area of rectangle is:",area_of_rectangle)
